requirements.txt
pandas~=1.5.3
dask~=2022.12.1
nextcord~=2.4.2
PyYAML~=6.0
aiohttp~=3.8.4
aiofiles~=23.1.0
numpy~=1.24.3
base58~=2.1.1
pip~=23.1.2
attrs~=22.2.0
distro~=1.8.0
wheel~=0.40.0
cryptography~=40.0.2
Pillow~=9.5.0
tornado~=6.3.1
Jinja2~=3.1.2
toolz~=0.12.0
partd~=1.4.0
fsspec~=2023.5.0
packaging~=23.1
cloudpickle~=2.2.1
click~=8.1.3
distributed~=2022.12.1
fastparquet~=2023.4.0
psutil~=5.9.5
tblib~=1.7.0
pytz~=2023.3
idna~=3.4
multidict~=6.0.4
zict~=3.0.0
Cython~=0.29.34
setuptools~=67.7.0
cffi~=1.15.1
locket~=1.0.0
six~=1.16.0
requests~=2.28.2
yarl~=1.9.2
MarkupSafe~=2.1.2
lxml~=4.9.2
python-dateutil~=2.8.2
aiosignal~=1.3.1
frozenlist~=1.3.3
cramjam~=2.6.2
Twisted~=22.10.0
pycurl~=7.45.2
pydantic~=1.10.7
pyparsing~=3.0.9
zipp~=3.15.0
msgpack~=1.0.5
setproctitle~=1.3.2
sortedcontainers~=2.4.0